+++ server/test/unit/SessionStateTest.js
@@ -8,23 +8,23 @@ const it = require('mocha').it
 const { sessionFactory } = require('../testingHelpers')
 
 describe('Session class', () => {
-  let state
+  let session
 
   beforeEach(() => {
-    state = sessionFactory({
-      numPresses: 1,
+    session = sessionFactory({
+      numButtonPresses: 1,
     })
   })
 
   it('should start off with 1 button press', () => {
-    expect(state).to.have.property('numPresses')
-    expect(state.numPresses).to.deep.equal(1)
+    expect(session).to.have.property('numButtonPresses')
+    expect(session.numButtonPresses).to.deep.equal(1)
   })
 
   it('should increment properly', () => {
-    state.incrementButtonPresses(1)
-    expect(state.numPresses).to.deep.equal(2)
-    state.incrementButtonPresses(2)
-    expect(state.numPresses).to.deep.equal(4)
+    session.incrementButtonPresses(1)
+    expect(session.numButtonPresses).to.deep.equal(2)
+    session.incrementButtonPresses(2)
+    expect(session.numButtonPresses).to.deep.equal(4)
   })
 })